requirements.txt
scipy==1.7.3
paddlepaddle==2.3.0
numpy==1.19.5
pandas==1.3.5
scikit_learn==0.24.1
